+++ service-task/src/main/java/com/github/hollykunge/security/task/entity/LarkSourceLink.java
@@ -3,8 +3,8 @@ package com.github.hollykunge.security.task.entity;
 import com.github.hollykunge.security.common.entity.BaseEntity;
 import lombok.Data;
 
-import java.util.Date;
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Table;
 
 /**
  * @author  fansq
@@ -34,6 +34,6 @@ public class LarkSourceLink extends BaseEntity {
     private String organizationCode;
 
     @Column(name = "SORT")
-    private Long sort;
+    private Integer sort;
 
 }
