+++ service-task/src/main/java/com/github/hollykunge/security/task/entity/LarkTaskStagesTemplate.java
@@ -3,8 +3,8 @@ package com.github.hollykunge.security.task.entity;
 import com.github.hollykunge.security.common.entity.BaseEntity;
 import lombok.Data;
 
-import java.util.Date;
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Table;
 
 /**
  * @author  fansq
@@ -22,7 +22,7 @@ public class LarkTaskStagesTemplate extends BaseEntity {
     private String projectTemplateCode;
 
     @Column(name = "SORT")
-    private Long sort;
+    private Integer sort;
 
     @Column(name = "CODE")
     private String code;
